+++ docs/changelog.rst
@@ -15,7 +15,7 @@ Changelog
 0.0.23 (2020-08-20)
 ~~~~~~~~~~~~~~~~~~~
 
--  :meth:`ocdsextensionregistry.profile_builder.ProfileBuilder.get_standard_file_contents`: Fix for OCDS 1.1.5.
+-  :class:`~ocdsextensionregistry.profile_builder.ProfileBuilder`: :meth:`~ocdsextensionregistry.profile_builder.ProfileBuilder.get_standard_file_contents`: Fix for OCDS 1.1.5.
 
 0.0.22 (2020-08-11)
 ~~~~~~~~~~~~~~~~~~~